Rename variables to be more descriptive
There were some variables that were placed in the project with
generic names such as `step1`, `step2`, and `stepN`.  These
should be replaced with more descriptive variable names.

Modify build_range_stream_rev behavior

The build_range_stream_rev behavior attempts to build ranges
in the given number of chunk_sizes.  It did this by starting at the
target and working backwards.  However, this leads to a couple
of concerns.  The first is that we may be missing a chunked
range due to not accounting for fractional iterations.  The
second is that we are working backwards with a check against
our decrement to ensure that it does not go below zero for
unsigned integers.  The unsigned check is errant and will not
prevent overflows.  In addition, because this returns a `Stream`
instead of an `Iterator`, it makes it somewhat less useful for
reuse.

This change builds these ranges from 0 forward, and then
delivers them in reverse order.  This should allow us to avoid any
overflow issues related to unsigned integers.  In addition, we
also ensure that we are accounting for all chunks by ensuring
that we take the ceiling of our division.  We also ensure that we
do not exceed our target by comparing our new end with the
passed in end and taking the minimum between them.  Finally
we modify this function to return an Iterator for better flexibility
and reusability.

Modify expand_range_to_block_stream generation

expand_range_to_block_stream is emitting  nested streams
requiring a `.flatten` call in order to resolve them.  The
preference is to resolve them with a `Future` by `await`ing
instead.

Remove unnecessary clone call

the local binding of block_data to data with a clone is
unnecessary, and as such should be removed entirely.

Modify Stream to keep Results

Instead of correcting the failures, it is desired to actually have
the Results of the Stream calls and transformations persist
throughout the accumulation process.  In addition to correcting
the function call to adhere to the modifications of the helper
functions, we also ensure that we keep a `Result` wrapping the
`Transaction` in the `Stream`.  This allows us to perform a
`try_collect` at the end, and report an error on any missing or
failed to retrieve / enumerate transactions.