Add tests for transaction/range/rev endpoint
In order to ensure that the new `/transaction/range/rev`
endpoint behaving as expected, a test is needed in order to
ensure that the correct transactions are being returned, in the
correct order.

This commit adds a test to ensure that the correct number of
transactions are being returned, and in the correct order.