Cleanup docker containers on termination
The `--rm` switch tells docker to cleanup containers on
termination. Without this an exlicit `remove` is required after
`stop` to clean up the container and a `prune` to remove arbitarary
volumes left behind. While these are not difficult actions to perform,
we are often unaware that we need to do them.

I've also changed `kill` to `stop` because `stop` in the drop fn
because `stop` attempts to shutdown the container gracefully.