Add Errors, Types, and Trait definitions for Explorer API
In order to implement an API that fulfills the needs of a Block Explorer
and in order to keep is separate from other priorities, it makes sense to
create a separate API module that doesn't require all of the same
guarantees and behaviors that the other modules require.

This commit servers as an initial definition of the endpoints with no
stubbed behavior to serve as a foundation for implementations.  This
commit adds some initial specialized Errors, and type definitions that
serve to illustrate the desired design and behavior.  Some types have
fields omitted at the moment, as the type pollution is potentially
difficult to reconcile.

In addition the storage and data sources have been modified to
implement this new trait with stub implementations that indicate that
they have not been implemented.