Add get_transaction_detail for ExplorerStorage
In order for the endpoint get_transaction_detail to be able to return the relevant
response for the requested transaction, we need to be able to retrieve the relevant
data from the identified source.  As such, the ExplorerStorage needs to be able to
retrieve the stored data.

This commit adds the requirement of the method get_transaction_detail to
ExplorerStorage as well as provides an initial implementation for the SQL storage
layer.  The other storage types have stubbed implementations for now.