Implement path parameter parsing for explorer api
The explorer api returned by define_api currently doesn't consider any of the
potential arguments that are being passed to the endpoint.  In order for this
API to be able to return paginated results correctly we need the ability for these
parameters to modify the paging behavior.

This commit implements these endpoints, and also indicates to the user
whether or not the passed limit parameter is valid or not (when relevant).

Add error for invalid limit parameter

When a passed limit is invalid, we need to be able to indicate this to the
user with a correctly formatted error message that fits the Explorer APIs
requirements.

This commit adds the InvalidLimit error and returns it where applicable.