Remove Duration for block_time
The std::time::Duration parameter for the explorer histograms
ends up causing extra parse burden on the block explorer,
furthermore it is encoded in a non-standard duration
representation.  It is best to just return the raw integer for
duration for now, and if we do want it to be more specific
we could potentially utilize ISO8601 for formatting durations.

This commit replaces the std::time::Duration for block time
with u64 instead.

Perhaps this may need to consider signed integers?  Is it
possible for a block time to be negative?  For now we assume
that it isn't.