Implement ExplorerDataSource for ExtensibleDataSource
The ExtensibleDataSource is the primary hopping on point for the espresso-sequencer.
As such, any data source that is going to be made available in the espresso-sequencer
needs to be implemented for the ExtensibleDataSource.

This commit implements the ExplorerDataSource for the ExtensibleDataSource by
passing all calls through to the underlying DataSource.