Fix query for pulling block summaries
When retrieving block summaries from the database we use a start point and retrieve
that starting point and prior blocks up to our limit.  There is an error in the query when
we retrieve block summaries by height, it constrains the considered blocks to only a
single block instead of that block and all prior.

This is fixed by utilizing the correct comparison operation on the query.