Fix: TmpDb::init() wait for psql
Currently `TmpDb::init()` attempts to wait for postgres to become ready,
but the logic will panic if `psql` is not available on the
host. But, we should not have any expectation that `psql` be available
on the host. This PR updates the wait loop to run a query w/ `psql` inside
the docker container.

NOTE due to incoming hotshot and sequencer changes this hasn't been properly tested so there
may be some bugs. Once hotshot is updated we can test this properly.