Refactor database transactions
A DB transaction is a proper type now in its own right, as opposed to having
transaction logic (commit, revert) as functions on the main server object, which
lead to some horrific ownership semantics and footguns (like the need to
explicitly revert a transaction if it failed, or else all future modifications
would use the same aborted transaction and also fail). The new design is much
cleaner and also a better match for SQLx's ownership semantics. From here on
out, actually switching from tokio-postgres to sqlx should be straightforward.

One immediate consequence of this change is that all mutable operations now
happen on the transaction, rather than the data source itself, so we no longer
need to acquire a write lock to mutate the database. This should be a huge
performance win as it allows reads to proceed while an update is in flight,
avoiding long queues on the RwLock. It also allows concurrent writers to the
extent that the underlying database does. Of course, we still have a single
connection, so we won't see the benefits of this immediately, but the benefits
will be huge once we switch to sqlx and use their connection pool.