Remove unused Default bound on InstanceState
In the sequencer repo we don't have a proper `Default` implentation but
one that's gated behind the testing feature because it wouldn't actually
work.

Working on
https://github.com/EspressoSystems/espresso-sequencer/pull/2283
it seems that this is one of the pain points for removing the spread of
the testing features in the sequencer crates.

It looks like we don't need the trait bound anymore in the query
service.